+++ NEWS.md
@@ -7,6 +7,7 @@ editor_options:
 - Replaces 55 HF items with 48 HF items in `builtin_itemtable` to reflect dropping of the first three months form
 - Updates `sample_hf` example data to the 48-item HF version
 - Adds `hf_48_2406.txt` and `hf_48_2510.txt` key files
+- Updates `builtin_itembank` using 48-item version of `gh1` for keys `gsed2212`, `gsed2404` and `gsed2510`.
 
 # dscore 2.0.3